+++ ios/Classes/Modules/BugReportingApi.m
@@ -146,4 +146,23 @@ extern void InitBugReportingApi(id<FlutterBinaryMessenger> messenger) {
     };
 }
 
+- (void)setDisclaimerTextText:(NSString *)text error:(FlutterError *_Nullable *_Nonnull)error {
+    [IBGBugReporting setDisclaimerText:text];
+}
+
+- (void)setCommentMinimumCharacterCountLimit:(NSNumber *)limit reportTypes:(nullable NSArray<NSString *> *)reportTypes error:(FlutterError *_Nullable *_Nonnull)error {
+    IBGBugReportingReportType resolvedTypes = 0;
+
+    if (![reportTypes count]) {
+        resolvedTypes = (ArgsRegistry.reportTypes[@"ReportType.bug"]).integerValue | (ArgsRegistry.reportTypes[@"ReportType.feedback"]).integerValue | (ArgsRegistry.reportTypes[@"ReportType.question"]).integerValue;
+    }
+    else {
+        for (NSString *reportType in reportTypes) {
+            resolvedTypes |= (ArgsRegistry.reportTypes[reportType]).integerValue;
+        }
+    }
+    
+    [IBGBugReporting setCommentMinimumCharacterCountForReportTypes:resolvedTypes withLimit:limit.intValue];
+}
+
 @end