+++ pigeons/bug_reporting.api.dart
@@ -27,4 +27,9 @@ abstract class BugReportingHostApi {
   );
   void bindOnInvokeCallback();
   void bindOnDismissCallback();
+  void setDisclaimerText(String text);
+  void setCommentMinimumCharacterCount(
+    int limit,
+    List<String>? reportTypes,
+  );
 }