+++ test/bug_reporting_test.dart
@@ -178,4 +178,25 @@ void main() {
       mHost.bindOnDismissCallback(),
     ).called(1);
   });
+
+  test('[setDisclaimerText] should call host method', () async {
+    const text = 'This is a disclaimer text!';
+
+    await BugReporting.setDisclaimerText(text);
+
+    verify(
+      mHost.setDisclaimerText(text),
+    ).called(1);
+  });
+
+  test('[setCommentMinimumCharacterCount] should call host method', () async {
+    const count = 20;
+    const reportTypes = [ReportType.bug];
+
+    await BugReporting.setCommentMinimumCharacterCount(count, reportTypes);
+
+    verify(
+      mHost.setCommentMinimumCharacterCount(count, reportTypes.mapToString()),
+    ).called(1);
+  });
 }